Peer Evaluation App Server

Generates the server part of the peer evaluation app

peer_evaluation_server(
  roster,
  data_gs_title,
  gs_key_file,
  welcome_md_file,
  self_eval_plus_md_file,
  self_eval_minus_md_file,
  teammate_eval_plus_md_file,
  teammate_eval_minus_md_file,
  quant_scores_md_file,
  points_per_teammate = 10,
  max_points = 15,
  min_points = 0,
  min_point_difference = 2,
  require_self_eval = TRUE,
  require_peer_evals = TRUE,
  auto_login_access_code = NULL
)

Arguments

roster

data frame with the student roster

data_gs_title

name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the gs_key_file must grant access to it.

gs_key_file

path to your .json access key file for peer evaluation google spreadsheets. See the peer evaluations vignette for details. This key file is safe to use on a secure shiny app server but be careful to never post this file anywhere publicly as it could be used to gain access to your peer evaluation spreadsheets. Make sure to share the google spreadshhet for this peer evaluation with the client_email listed in the key file.

welcome_md_file

path to a markdown (.md) file for the login welcome message

self_eval_plus_md_file

markdown file for the "plus" self evaluation message

self_eval_minus_md_file

markdown file for the "minus" self evaluation message

teammate_eval_plus_md_file

markdown file for the "plus" teammate evaluation message

teammate_eval_minus_md_file

markdown file for the "minus" teammate evaluation message

quant_scores_md_file

markdown file for the quantiative scores message

points_per_teammate

points per teammate

max_points

the maximum number of points allowed per team member

min_points

the smallest number of points allowed per team member

min_point_difference

the minimum point difference required for the scores (set to 0 to allow all scores to be identical)

require_self_eval

whether the qualitative self evaluation is required

require_peer_evals

whether the qualitative self evaluation is required

auto_login_access_code

set an automatic login access code for testing and debugging purposes