tabsort

Returns a sorted (descending) frequency tbl

tabsort(.data, ..., prop = TRUE, na_omit = TRUE, sort = TRUE)

ntbl(.data, ...)

Arguments

.data
Data

...
Unquoted column names of variables to include in table. Default is to use all columns.

prop
Logical indicating whether to include a proportion of total obs column.

na_omit
Logical indicating whether to exclude missing. If all responses are missing, a missing value is used as the single category.

sort
Logical indicating whether to sort the returned object.

Value

Frequency tbl

Examples

## generate example data
x <- sample(letters[1:4], 200, replace = TRUE)
y <- sample(letters[5:8], 200, replace = TRUE)

## count and sort frequencies for each vector
tabsort(x)

#> # A tibble: 4 x 3
#>   x         n  prop
#>   <chr> <int> <dbl>
#> 1 b        55 0.275
#> 2 c        51 0.255
#> 3 a        49 0.245
#> 4 d        45 0.225

tabsort(y)

#> # A tibble: 4 x 3
#>   x         n  prop
#>   <chr> <int> <dbl>
#> 1 g        57 0.285
#> 2 f        49 0.245
#> 3 e        48 0.24 
#> 4 h        46 0.23 

## combine x and y into data frame
dat <- data.frame(x, y)

## select columns and create freq table
tabsort(dat, x)

#> # A tibble: 4 x 3
#>   x         n  prop
#>   <chr> <int> <dbl>
#> 1 b        55 0.275
#> 2 c        51 0.255
#> 3 a        49 0.245
#> 4 d        45 0.225

tabsort(dat, x, y)

#> # A tibble: 16 x 4
#>    x     y         n  prop
#>    <chr> <chr> <int> <dbl>
#>  1 c     g        18 0.09 
#>  2 a     f        15 0.075
#>  3 d     g        15 0.075
#>  4 b     h        15 0.075
#>  5 b     g        14 0.07 
#>  6 b     e        13 0.065
#>  7 d     e        13 0.065
#>  8 b     f        13 0.065
#>  9 c     f        13 0.065
#> 10 a     h        13 0.065
#> # … with 6 more rows