Arrange RAT questions

Function to make it easy to arrange an RAT either by fixed numbers (fixed_number_column) or semi-randomly or completely randomly. Can also arrange questions within specific groups using the group_by_column, and can use specific starting numbers for iRAT and tRAT numbering (which need not be the same in case tRATs need to match a specific IF-AT number sequence and iRATs should match a scantron). In case iRAT numbers are different from tRAT numbers, the tRAT numbering is the one matched to the answer key since this needs to match the IF-AT scratch-off.

tbl_arrange_RAT_questions(rat, by = "original", tRAT_n_start = 1,
  iRAT_n_start = tRAT_n_start, iRAT_sel_per_q = 1,
  fixed_number_column = NULL, group_by_column = NULL,
  random_seed = random())

Arguments

rat
Readiness Assessment Test object

by
what to arrange by, options: by="original" leaves the original order as encountered, by="random" generates a random order, by="semi-random" generates a random order except for questions that have the fixed_number_column set, and by="fixed" uses the fixed_number_column (which has to be set for all questions in the latter case!).

tRAT_n_start
the start number for the tRAT question numbering (if the rRAT should fit a different part of an IF-AT), only matters if by is NOT "fixed". Default is 1.

iRAT_n_start
the start number for the iRAT question numbering. By default the same as tRAT_n_start. Specify a different value from tRAT_n_start e.g. if tRAT starts at a higher IF-AT number but iRAT numbers should start at 1.

iRAT_sel_per_q
number of selections per question for the iRAT portion of this test, e.g. when using scantrons and giving the students the option to hedge their bets on the iRAT portion of the test.

fixed_number_column
name of a column in the RAT questions data frame that indicates the fixed question number. Only relevant if by="semi-random" or by="fixed"

group_by_column
name of a column in the RAT questions data frame that indicates which questions to group together (groups will be arranged alphabetically, those with undefined group come last). Only relevant if by="random" or by="semi-random". Throws an error if grouping and any fixed number questions are incompatible (e.g. a question has fixed number 5 but is part of the first group of only 3 questions).

random_seed
can overwrite with a fixed value (e.g. random_seed=42) to get a reproducible "random" order in by="random" or by="semi-random" mode.

Value

returns the RAT object with the location in the test specified