Create RAT from Excel

Load RAT questions and answer key from Excel. Note that logical columns ('correct' and 'include') are evaluated to TRUE if they are =1, =TRUE, ='TRUE', ='yes' or ='x', everything else is FALSE. Uses tbl_create_RAT_from_data_frame internally.

tbl_create_RAT_from_excel(filepath, questions_tab = "questions",
  key_tab = "key", filter_include = TRUE, fill_down_questions = TRUE)

Arguments

filepath
the path to the excel file

questions_tab
the name of the questions tab (requires at minimum columns 'question', 'answer', and logical TRUE/FALSE 'correct', plus logical 'include' if filter_include=TRUE)

key_tab
the name of th keys tab (requires at mimimum columns 'number', 'option')

filter_include
if set, only keeps questions that have the 'include' column set

fill_down_questions
whether to fill down the questions column (i.e. if question an their parameters are only written in first row)