Retrieve default RAT layouts

This function provides the default layout options for TBL questions (horizontal and vertical arrangement) but can easily be expanded with custom layouts. To introduce additional/alternative question layouts, simply overwrite the answer_layout_funs parameter in tbl_generate_RAT_choices with a list of functions that have the layout names as keys and functions that take two parameters (a vector of answer_option letters A, B, C, D, etc. and an answer vector of the same length with the actual answers) as values. The functions must return valid markdown.

tbl_default_RAT_layouts()