Duplicate a peer evaluation app

This functions makes it easy to duplicate an existing peer evaluation by copying all relevant files into a new folder. You can optionally change the google spreadsheet the new peer evaluation is linked to by specifying the data_gs_title parameter. To create a new peer evaluation from scratch, use tbl_setup_peer_evaluation instead.

tbl_duplicate_peer_evaluation(
  template = "peer_evaluation",
  folder = "peer_evaluation2",
  data_gs_title = NULL,
  overwrite = FALSE,
  check_gs_access = TRUE
)

Arguments

template
folder where the existing peer evaluation that is to be duplicated is located (path must be either relative to the current working directory or an absolute file path on the operating system).

folder
target folder where to setup the duplicated peer evaluation (path must be either relative to the current working directory or an absolute file path on the operating system). If the folder does not exist yet, it will be created automatically.

data_gs_title
name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the gs_key_file must grant access to it.

overwrite
whether to overwrite the app in the target directory if it already exists

check_gs_access
whether to confirm google spreadsheet access (using the tbl_check_gs_access function). Note that if this is set to FALSE, this function will NOT validate the gs_key_file and NOT check that the provided data_gs_title is a valid spreadsheet the key file grants access to.

Value

returns the folder invisibly for ease of use in pipelines

See also