Fetch the immediate feedback test data

Fetches the app data from the google spreadsheet and reads it (using tbl_read_immediate_feedback_test_data). For standard installations of the immediate feedback test app, all defaults should be sufficient.

tbl_fetch_immediate_feedback_test_data(
  folder = ".",
  data_gs_title = get_app_data_gs_title(folder),
  data_gs_id = NULL,
  roster = read_excel(file.path(folder, "roster.xlsx")),
  questions = suppressMessages(read_excel(file.path(folder, "questions.xlsx"))),
  gs_key_file = file.path(folder, "gs_key_file.json"),
  download_to = file.path(folder, "immediate_feedback_test_data_downloaded.xlsx")
)

Arguments

folder

folder where the immediate feedback test app is located (relative to the location of the RMarkdown file if used in the latter context)

data_gs_title

name of the google spreadsheet that should be used for storing the data for the immediate feedback test. This spreadsheet must already exist and the gs_key_file must grant access to it.

data_gs_id

optional alternative to the data_gs_title, a google spread sheet ID text or object (see sheets_id). If provided, it takes precedence over the data_gs_title parameter as the google spreadsheet doesn't need to be searched by name anymore (i.e. can be loaded faster).

roster

data frame with the roster and access codes

questions

data frame with the questions

gs_key_file

path to your .json access key file for TBL google spreadsheets. See the peer evaluations vignette for details. This key file is safe to use on a secure shiny app server but be careful to never post this file anywhere publicly as it could be used to gain access to your TBL spreadsheets. Make sure to share the google spreadshhet for this immediate feedback test with the client_email listed in the key file.

download_to

location where the whole immediate feedback test data sheet will be downloaded to for more efficient read access