Run a peer evaluation app

This function starts the peer evaluation user interface. Note that this function is typically NOT called directly but indirectly by setting up the peer evalution app using tbl_setup_peer_evaluation, adjusting the files in the peer evaluation app folder, and test running the app using tbl_test_peer_evaluation.

tbl_run_peer_evaluation(
  data_gs_title,
  roster = read_excel("roster.xlsx"),
  app_title = "Peer Evaluation",
  welcome_md_file = "app_welcome.md",
  self_eval_plus_md_file = "app_self_evaluation_plus.md",
  self_eval_minus_md_file = "app_self_evaluation_minus.md",
  teammate_eval_plus_md_file = "app_teammate_evaluation_plus.md",
  teammate_eval_minus_md_file = "app_teammate_evaluation_minus.md",
  quant_scores_md_file = "app_quantitative_scores.md",
  points_per_teammate = 10,
  max_points = 15,
  min_points = 0,
  min_point_difference = 2,
  require_self_eval = TRUE,
  require_peer_evals = TRUE,
  auto_login_access_code = NULL,
  ...,
  launch = FALSE
)

Arguments

data_gs_title

name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the gs_key_file must grant access to it.

roster

data frame with the student roster

app_title

the title of the Peer Evaluation App, e.g. "Class name - Peer Evaluation #1"

welcome_md_file

path to a markdown (.md) file for the login welcome message

self_eval_plus_md_file

markdown file for the "plus" self evaluation message

self_eval_minus_md_file

markdown file for the "minus" self evaluation message

teammate_eval_plus_md_file

markdown file for the "plus" teammate evaluation message

teammate_eval_minus_md_file

markdown file for the "minus" teammate evaluation message

quant_scores_md_file

markdown file for the quantiative scores message

points_per_teammate

points per teammate

max_points

the maximum number of points allowed per team member

min_points

the smallest number of points allowed per team member

min_point_difference

the minimum point difference required for the scores (set to 0 to allow all scores to be identical)

require_self_eval

whether the qualitative self evaluation is required

require_peer_evals

whether the qualitative self evaluation is required

auto_login_access_code

set an automatic login access code for testing and debugging purposes

...

passed on to the runApp call (only if launch = TRUE), can include server-specific parameters such as host or port

launch

whether to launch the app (TRUE) or return a shiny app object (FALSE) that then can be launched via runApp

See also