Set up a peer evaluation app

This functions makes it easy to setup a peer evaluation app folder. It creates the app script app.R and copies the roster_file (an Excel spreadsheet) as well as the necessary google drive credentials for the data_gsheet to the same directory. It will ask for google drive credentials using gs_auth and the entered credentials must have access to the data_gsheet. Both app.R and the roster_file can be edited and tested in the app directory before uploading the whole app to a shiny server. On the shiny server, the roster_file will be read only and all actual data will be stored in the google drive data file.

tbl_setup_peer_evaluation(folder = "peer_evaluation",
  data_gs_title = "Peer Evaluation",
  template_roster_file = system.file(package = "tbltools", "extdata",
  "roster_template.xlsx"), gs_token = file.path(folder, "gs_token.rds"),
  overwrite = FALSE)

Arguments

folder
target folder where to setup the peer evaluation (path must be either relative to the current working directory or an absolute file path on the operating system). If the folder does not exist yet, it will be created automatically.

data_gs_title
name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the credentials used when asked by this function must have write access to the spreadsheet.

template_roster_file
path to an excel (.xslx) file that contains the student roster information to use as template for the peer evaluation app (can be edited later) - will use the package template by default

gs_token
path to a google spreadsheet oauth 2.0 authentication token file (see Token-class). If none is provided or the file does not exist yet, will ask for google drive credentials interactively to generate a token for the peer evaluation app. The token is safe to use on a secure shiny app server but be careful to never post this token file anywhere publicly as it could be used to gain access to your google drive.

overwrite
whether to overwrite the app in the target directory if it already exists

Value

returns the folder invisibly for ease of use in pipelines

Details

Use tbl_test_peer_evaluation with the same folder as parameter to test run the peer evaluation application locally on your computer.

See also

Other peer evaluation functions: tbl_run_peer_evaluation, tbl_test_peer_evaluation