Test run peer evaluation locally

This function starts a peer evaluation app previously set up using tbl_setup_peer_evaluation. This provides a local version of the app before deploying it to a shiny apps server (using tbl_deploy_peer_evaluation). Keep in mind that the test application writes data to the google spreadsheet and make sure to delete any test records from the google spreadsheet that may interfere with any students' peer evaluations prior to providing the students with the link to the deployed peer evaluation app.

tbl_test_peer_evaluation(folder = "peer_evaluation", ...)

Arguments

folder

folder where the peer evaluation app is located (see tbl_setup_peer_evaluation for details)

...

optional parameters passed to runApp

See also