Peer Evaluation App Server

Generates the server part of the peer evaluation app

peer_evaluation_server(roster, data_gs_title, gs_token, welcome_md_file,
  self_eval_plus_md_file, self_eval_minus_md_file, teammate_eval_plus_md_file,
  teammate_eval_minus_md_file, quant_scores_md_file, points_per_teammate = 10,
  max_points = 15, min_points = 0, min_point_difference = 2,
  auto_login_access_code = NULL)

Arguments

roster
data frame with the student roster

data_gs_title
name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the credentials used when asked by this function must have write access to the spreadsheet.

gs_token
path to a google spreadsheet oauth 2.0 authentication token file (see Token-class). If none is provided or the file does not exist yet, will ask for google drive credentials interactively to generate a token for the peer evaluation app. The token is safe to use on a secure shiny app server but be careful to never post this token file anywhere publicly as it could be used to gain access to your google drive.

welcome_md_file
path to a markdown (.md) file for the login welcome message

self_eval_plus_md_file
markdown file for the "plus" self evaluation message

self_eval_minus_md_file
markdown file for the "minus" self evaluation message

teammate_eval_plus_md_file
markdown file for the "plus" teammate evaluation message

teammate_eval_minus_md_file
markdown file for the "minus" teammate evaluation message

quant_scores_md_file
markdown file for the quantiative scores message

points_per_teammate
points per teammate

max_points
the maximum number of points allowed per team member

min_points
the smallest number of points allowed per team member

min_point_difference
the minimum point difference required for the scores (set to 0 to allow all scores to be identical)

auto_login_access_code
set an automatic login access code for testing and debugging purposes