Set up a peer evaluation app

This functions makes it easy to setup a peer evaluation app folder. It creates the app script app.R and copies the roster_file (an Excel spreadsheet) as well as the necessary google access key file for the data_gs_title google spreadsheet to the same directory. The provided gs_key_file must grant access to the google spreadsheet (data_gs_title). See the peer evaluations vignette for details on how to set up google credentials and generate a key file. Unless check_gs_access = FALSE, this function will call the tbl_check_gs_access function to confirm that the key file works and provides access to the requested google spreadsheet. Both app.R and the roster_file can be edited and tested in the app directory before uploading the whole app to a shiny server. On the shiny server, the roster_file will be read-only and all actual data will be stored in the google spreadsheet data file.

tbl_setup_peer_evaluation(
  folder = "peer_evaluation",
  data_gs_title = "Peer Evaluation",
  template_roster_file = system.file(package = "tbltools", "extdata",
    "roster_template.xlsx"),
  gs_key_file = NULL,
  gs_token = deprecated(),
  overwrite = FALSE,
  check_gs_access = TRUE
)

Arguments

folder

target folder where to setup the peer evaluation (path must be either relative to the current working directory or an absolute file path on the operating system). If the folder does not exist yet, it will be created automatically.

data_gs_title

name of the google spreadsheet that should be used for storing the peer evaluation data. This spreadsheet must already exist and the gs_key_file must grant access to it.

template_roster_file

path to an excel (.xslx) file that contains the student roster information to use as template for the peer evaluation app (can be edited later) - will use the package template by default

gs_key_file

path to your .json access key file for peer evaluation google spreadsheets. See the peer evaluations vignette for details. This key file is safe to use on a secure shiny app server but be careful to never post this file anywhere publicly as it could be used to gain access to your peer evaluation spreadsheets. Make sure to share the google spreadshhet for this peer evaluation with the client_email listed in the key file.

gs_token

deprecated in favor of using a gs_key_file

overwrite

whether to overwrite the app in the target directory if it already exists

check_gs_access

whether to confirm google spreadsheet access (using the tbl_check_gs_access function). Note that if this is set to FALSE, this function will NOT validate the gs_key_file and NOT check that the provided data_gs_title is a valid spreadsheet the key file grants access to.

Value

returns the folder invisibly for ease of use in pipelines

Details

Use tbl_test_peer_evaluation with the same folder as parameter to test run the peer evaluation application locally on your computer. It is recommended to always test the peer evaluation application locally before uploading it to the shiny application server.

Use tbl_duplicate_peer_evaluation to create a copy of an existing peer evaluation rather than setting one up from scratch. Note that older peer evaluations that don't have a gs_key_file yet will require this parameter to be set.

See also